requirements.txt
# Web Framework
streamlit>=1.28.0

# GraphQL API
gql[all]>=3.4.0
aiohttp>=3.9.0
graphql-core>=3.2.0

# Image Processing
large-image[all]>=1.25.0
openslide-python>=1.3.0
tifffile>=2023.7.0
scikit-image>=0.21.0
numpy>=1.24.0,<2.0.0
pillow>=10.0.0
opencv-python>=4.8.0

# ML Models
torch>=2.0.0
torchvision>=0.15.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
